+++ Atelier Meruru DX/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2024/10/22
+1. added alchemy exp multiplier
+1. added set game window to borderless mode
+
 ## 2024/09/16
 1. minor tweaks